Feat: slider
