+++ src/dataPatch/fetchAndStoreBreeds.jsx
@@ -1,7 +1,10 @@
 // src/dataPatch/fetchAndStoreBreeds.js
-import { collection, getDocs } from 'firebase/firestore';
-import { db } from '../firebase';
 
+import { collection, getDocs } from 'firebase/firestore';
+import { ref, getDownloadURL, listAll } from 'firebase/storage';
+import { db, storage } from '../firebase';
+
+// Fetch breeds data from Firestore
 const fetchBreedsFromFirestore = async () => {
   try {
     const querySnapshot = await getDocs(collection(db, 'dogs'));
@@ -12,15 +15,14 @@ const fetchBreedsFromFirestore = async () => {
   }
 };
 
+// Merge breeds data and sort by Korean names
 const mergeBreedsData = async (firestoreData) => {
-  // Sorting the data by Korean names in alphabetical order
   const sortedData = firestoreData.sort((a, b) => {
     const nameA = a.koreanName || '';
     const nameB = b.koreanName || '';
     return nameA.localeCompare(nameB, 'ko');
   });
 
-  // Merging the sorted data
   const mergedData = sortedData.reduce((acc, breed) => {
     acc[breed.englishName.toLowerCase()] = breed;
     return acc;
@@ -29,16 +31,41 @@ const mergeBreedsData = async (firestoreData) => {
   return mergedData;
 };
 
+// Fetch images from Firebase Storage
+const fetchImagesFromStorage = async (breedName) => {
+  try {
+    const formattedBreedName = breedName.replace(/ /g, '_');
+    const folderRef = ref(storage, `dog/${formattedBreedName}`);
+    const fileList = await listAll(folderRef);
+
+    if (fileList.items.length > 0) {
+      const imageUrls = await Promise.all(
+        fileList.items.map(async (itemRef) => {
+          const url = await getDownloadURL(itemRef);
+          return url;
+        })
+      );
+      return imageUrls;
+    } else {
+      return [];
+    }
+  } catch (error) {
+    console.error(`Error fetching images from Storage for breed ${breedName}:`, error);
+    return [];
+  }
+};
+
+// Fetch and store breeds data in localStorage
 const fetchAndStoreBreeds = async () => {
   const firestoreData = await fetchBreedsFromFirestore();
   const mergedData = await mergeBreedsData(firestoreData);
-  
   localStorage.setItem('breedsData', JSON.stringify(mergedData));
 };
 
+// Get breeds data from localStorage
 const getBreedsData = () => {
   const data = localStorage.getItem('breedsData');
   return data ? JSON.parse(data) : null;
 };
 
-export { fetchAndStoreBreeds, getBreedsData };
+export { fetchAndStoreBreeds, getBreedsData, fetchImagesFromStorage };
